Release 1.10:     - The plugin is updated to use Parallel Colt 0.8.2.
git-svn-id: http://spectraldeconv.svn.sourceforge.net/svnroot/spectraldeconv@3 aaff9079-7fab-494d-bae9-a53d918c268b

+++ ParallelSpectralDeconvolution/src/edu/emory/mathcs/restoretools/spectral/tsvd/DoubleReflexiveTruncatedSVD2D.java
@@ -155,7 +155,7 @@ public class DoubleReflexiveTruncatedSVD2D extends AbstractDoubleSpectralDeconvo
         int[] indices = DoubleSorting.quickSort.sortIndex(s, compDec);
         s = s.viewSelection(indices);
         bhat = bhat.viewSelection(indices);
-        int n = s.size();
+        int n = (int)s.size();
         double[] rho = new double[n - 1];
         rho[n - 2] = bhat.getQuick(n - 1) * bhat.getQuick(n - 1);
         DoubleMatrix1D G = new DenseDoubleMatrix1D(n - 1);
